rag-mcp/requirements.txt
mcp>=1.0.0
chromadb>=0.4.0
llama-cpp-python>=0.2.0
rank-bm25>=0.2.2
konlpy>=0.6.0
scikit-learn>=1.3.0
watchfiles>=0.21.0
# Text Extraction Libraries
pymupdf4llm>=0.0.12
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
polars>=0.20.0
python-magic>=0.4.27
markitdown>=0.0.1